add events

+++ insurance-fund/programs/insurance-fund/src/states/admin.rs
@@ -15,7 +15,7 @@ pub enum Permissions {
 
 #[account]
 pub struct Admin {
-    pub index: u64,
+    pub index: u8,
     pub address: Pubkey,
     pub permissions: Permissions
 }
